cleanup fit result

+++ pytesdaq/instruments/control.py
@@ -1010,12 +1010,12 @@ class Control:
         else:
 
             # signal generator parameter
-            output_dict['voltage'] = self._signal_generator_inst.get_amplitude(source=signal_gen_num)
+            output_dict['voltage'] = float(self._signal_generator_inst.get_amplitude(source=signal_gen_num))
             resistance = float(self._config.get_signal_gen_tes_resistance())
-            output_dict['current']  = output_dict['voltage']/resistance
-            output_dict['frequency'] = self._signal_generator_inst.get_frequency(source=signal_gen_num)
+            output_dict['current']  = float(output_dict['voltage']/resistance)
+            output_dict['frequency'] = float(self._signal_generator_inst.get_frequency(source=signal_gen_num))
             output_dict['shape'] = self._signal_generator_inst.get_shape(source=signal_gen_num)
-
+            
             # source
             if self._squid_controller_name == 'feb':
